Split LinuxGLContext into separate classes for X11 and Wayland
