+++ sources/Renderer/OpenGL/Platform/Linux/LinuxGLContext.h
@@ -1,149 +1,25 @@
 /*
  * LinuxGLContext.h
  *
- * Copyright (c) 2015 Lukas Hermanns. All rights reserved.
+ * Copyright (c) 2025 Lukas Hermanns. All rights reserved.
  * Licensed under the terms of the BSD 3-Clause license (see LICENSE.txt).
  */
 
 #ifndef LLGL_LINUX_GL_CONTEXT_H
 #define LLGL_LINUX_GL_CONTEXT_H
 
-
 #include "../GLContext.h"
-#include "../../OpenGL.h"
-#include <LLGL/RendererConfiguration.h>
-#include <LLGL/Platform/NativeHandle.h>
-#include <X11/Xlib.h>
 
 
 namespace LLGL
 {
 
-enum class LinuxGLAPIType : char {
-    GLX,
-    EGL
-};
-
 // Implementation of the <GLContext> interface for GNU/Linux and wrapper for a native GLX context.
 class LinuxGLContext : public GLContext
 {
 
-    public:
-
-        LinuxGLContext(
-            const GLPixelFormat&                    pixelFormat,
-            const RendererConfigurationOpenGL&      profile,
-            Surface&                                surface,
-            LinuxGLContext*                         sharedContext,
-            const OpenGL::RenderSystemNativeHandle* customNativeHandle
-        );
-        ~LinuxGLContext();
-
-        int GetSamples() const override;
-
-        bool GetNativeHandle(void* nativeHandle, std::size_t nativeHandleSize) const override;
-
-    public:
-
-        // Tries to find an X11 visual configuration for the specified pixel format and
-        // modifies the sample count depending on availability. Returns null if no such visual could be found.
-        static ::XVisualInfo* ChooseVisual(::Display* display, int screen, const GLPixelFormat& pixelFormat, int& outSamples);
-
-        // Returns the native X11 <GLXContext> object.
-        inline void* GetGLXContext() const
-        {
-            return api_.glx.context;
-        }
-
-        inline bool IsEGL() const
-        {
-            return api_.type == LinuxGLAPIType::EGL;
-        }
-
-        #if LLGL_LINUX_ENABLE_WAYLAND
-
-        inline EGLConfig GetEGLConfig() const
-        {
-            return api_.egl.config;
-        }
-
-        #endif
-
-    private:
-
-        bool SetSwapInterval(int interval) override;
-
-    private:
-
-        void CreateGLXContext(
-            const GLPixelFormat&                pixelFormat,
-            const RendererConfigurationOpenGL&  profile,
-            const NativeHandle&                 nativeHandle,
-            LinuxGLContext*                     sharedContext
-        );
-
-        void DeleteGLXContext();
-
-        GLXContext CreateGLXContextCoreProfile(GLXContext glcShared, int major, int minor, int depthBits, int stencilBits);
-        GLXContext CreateGLXContextCompatibilityProfile(XVisualInfo* visual, GLXContext glcShared);
-
-        void CreateProxyGLXContext(
-            const GLPixelFormat&                    pixelFormat,
-            const NativeHandle&                     nativeWindowHandle,
-            const OpenGL::RenderSystemNativeHandle& nativeContextHandle
-        );
-
-        #if LLGL_LINUX_ENABLE_WAYLAND
-        EGLContext CreateEGLContextCoreProfile(EGLContext glcShared, int major, int minor, int depthBits, int stencilBits, EGLConfig* config);
-        EGLContext CreateEGLContextCompatibilityProfile(EGLContext glcShared, EGLConfig* config);
-
-        void CreateEGLContext(
-            const GLPixelFormat&                pixelFormat,
-            const RendererConfigurationOpenGL&  profile,
-            const NativeHandle&                 nativeHandle,
-            LinuxGLContext*                     sharedContext
-        );
-
-        void DeleteEGLContext();
-
-        void CreateProxyEGLContext(
-            const GLPixelFormat&                    pixelFormat,
-            const NativeHandle&                     nativeWindowHandle,
-            const OpenGL::RenderSystemNativeHandle& nativeContextHandle
-        );
-        #endif
-
-    private:
-        struct GLXData
-        {
-            ::Display*       display;
-            ::GLXContext     context;
-        };
-
-        #if LLGL_LINUX_ENABLE_WAYLAND
-        struct EGLData
-        {
-            EGLDisplay       display;
-            EGLContext       context;
-            EGLConfig        config;
-        };
-        #endif
-
-        struct ApiData
-        {
-            union
-            {
-                GLXData glx;
-                #if LLGL_LINUX_ENABLE_WAYLAND
-                EGLData egl;
-                #endif
-            };
-
-            LinuxGLAPIType type;
-        } api_;
-
-        int        samples_    = 1;
-        bool       isProxyGLC_ = false;
+public:
+    virtual bool IsWayland() const = 0;
 
 };
