+++ sources/Renderer/OpenGL/Platform/Linux/LinuxGLSwapChainContext.h
@@ -9,13 +9,10 @@
 #define LLGL_LINUX_GL_SWAP_CHAIN_CONTEXT_H
 
 
+#include "LinuxGLContextX11.h"
 #include "../GLSwapChainContext.h"
 #include "../../OpenGL.h"
 #include <X11/Xlib.h>
-
-#if LLGL_LINUX_ENABLE_WAYLAND
-#include <wayland-client.h>
-#endif
 
 namespace LLGL
 {
@@ -29,7 +26,7 @@ class LinuxX11GLSwapChainContext final : public GLSwapChainContext
 
     public:
 
-        LinuxX11GLSwapChainContext(LinuxGLContext& context, Surface& surface);
+        LinuxX11GLSwapChainContext(LinuxGLContextX11& context, Surface& surface);
 
         bool HasDrawable() const override;
         bool SwapBuffers() override;
@@ -47,14 +44,22 @@ class LinuxX11GLSwapChainContext final : public GLSwapChainContext
 
 };
 
+} // /namespace LLGL
+
+
 #if LLGL_LINUX_ENABLE_WAYLAND
+
+#include "LinuxGLContextWayland.h"
+#include <wayland-client.h>
+
+namespace LLGL {
 
 class LinuxWaylandGLSwapChainContext final : public GLSwapChainContext
 {
 
     public:
 
-        LinuxWaylandGLSwapChainContext(LinuxGLContext& context, Surface& surface);
+        LinuxWaylandGLSwapChainContext(LinuxGLContextWayland& context, Surface& surface);
 
         bool HasDrawable() const override;
         bool SwapBuffers() override;
@@ -71,10 +76,9 @@ class LinuxWaylandGLSwapChainContext final : public GLSwapChainContext
 
 };
 
-#endif
-
 } // /namespace LLGL
 
+#endif
 
 #endif
