+++ sources/Renderer/Vulkan/Ext/VKExtensionsDecl.inl
@@ -22,7 +22,10 @@ DECL_VKPROC( vkCreateWin32SurfaceKHR );
 #elif defined LLGL_OS_LINUX
 
 DECL_VKPROC( vkCreateXlibSurfaceKHR );
+
+#if LLGL_LINUX_ENABLE_WAYLAND
 DECL_VKPROC( vkCreateWaylandSurfaceKHR );
+#endif
 
 #elif defined LLGL_OS_ANDROID
